mqtt_pubsub/requirements-dev.txt
backports.tarfile==1.2.0
build==1.2.2.post1
certifi==2025.6.15
charset-normalizer==3.4.2
docutils==0.21.2
id==1.5.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.2.1
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
packaging==25.0
paho-mqtt==2.1.0
pluggy==1.6.0
py4j==0.10.9.9
pyarrow==20.0.0
Pygments==2.19.2
pyproject_hooks==1.2.0
pyspark==4.0.0
pytest==8.4.1
pytest-dotenv==0.5.2
python-dotenv==1.1.1
readme_renderer==44.0
requests==2.32.4
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
twine==6.1.0
urllib3==2.5.0
zipp==3.23.0
